Move plugin behavior to its own file
