+++ allennlp/common/testing/model_test_case.py
@@ -1,15 +1,15 @@
 import copy
-from typing import Any, Dict, Set, Union, Iterable
+from typing import Any, Dict, Iterable, Set, Union
 
-from numpy.testing import assert_allclose
 import torch
+from numpy.testing import assert_allclose
 
 from allennlp.commands.train import train_model_from_file
 from allennlp.common import Params
 from allennlp.common.testing.test_case import AllenNlpTestCase
 from allennlp.data import DataIterator, DatasetReader, Vocabulary
 from allennlp.data.batch import Batch
-from allennlp.models import Model, load_archive
+from allennlp.models import load_archive, Model
 
 
 class ModelTestCase(AllenNlpTestCase):
